Tried to add a list view for all users
Could not get it to work because onchangedata cannot add to list. Tried implementing callback but didnt work.

+++ app/src/main/java/project1/mobile/cs/fsu/edu/squad/SquadUser.java
@@ -63,11 +63,11 @@ public class SquadUser {
         return uPassword;
     }
 
-    public void setuLat(float uLat) {
+    public void setuLat(double uLat) {
         this.uLat = uLat;
     }
 
-    public void setuLonge(float uLonge) {
+    public void setuLonge(double uLonge) {
         this.uLonge = uLonge;
     }
